refactor storage module: rename storage to oss
