+++ pom.xml
@@ -81,7 +81,7 @@
         <module>spring-cloud-alibaba-test</module>
         <module>spring-cloud-starter-alibaba</module>
         <module>spring-cloud-starter-alicloud</module>
-        <module>spring-cloud-alicloud-storage</module>
+        <module>spring-cloud-alicloud-oss</module>
     </modules>
 
     <dependencyManagement>
