--- spring-cloud-starter-alicloud/spring-cloud-starter-alicloud-storage/pom.xml
@@ -1,24 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>org.springframework.cloud</groupId>
-        <artifactId>spring-cloud-starter-alicloud</artifactId>
-        <version>0.2.0.BUILD-SNAPSHOT</version>
-    </parent>
-    <artifactId>spring-cloud-starter-alicloud-storage</artifactId>
-    <name>Spring Cloud Starter Alibaba Cloud Storage</name>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.springframework.cloud</groupId>
-            <artifactId>spring-cloud-alicloud-storage</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.aliyun.oss</groupId>
-            <artifactId>aliyun-sdk-oss</artifactId>
-        </dependency>
-    </dependencies>
-
-</project>
